// src/app/models/aviso-cobranza/EstadoLote.ts
export enum EstadoLote {
  CARGADO = 'CARGADO',
  VALIDANDO = 'VALIDANDO',
  VALIDADO_OK = 'VALIDADO_OK',
  VALIDADO_CON_ERRORES = 'VALIDADO_CON_ERRORES',
  ERROR_VALIDACION = 'ERROR_VALIDACION',
  ENVIANDO = 'ENVIANDO',
  ENVIADO = 'ENVIADO',
  ERROR_ENVIO = 'ERROR_ENVIO',
  ELIMINADO = 'ELIMINADO',
}